first commit, code cleanup and start for test methods
