+++ .env
@@ -4,6 +4,7 @@ MONGO_INITDB_ROOT_PASSWORD=root
 MONGO_INITDB_DATABASE=db
 MONGO_USER=user
 MONGO_CONNECTION_STRING=mongodb://user:password@db:27017/db
+MONGO_CONNECTION_STRING_ADMIN=mongodb://root:root@db:27017/db
 RASA_ENDPOINT=http://rasa:5000
 RASA_PORT=5005
 REACT_APP_PORT=3000
