+++ server/Api/db_api.js
@@ -2,7 +2,7 @@
     Api endpoints which store or retrieve data from the database.
 */
 
-const db = require("../database")
+const db = require("../database");
 
 /*
     Post request. Params required :
@@ -322,157 +322,3 @@ exports.remove_entity = async function (req, res) {
     let status = await db.EntitiesModel.deleteOne({ 'name': params.name });
     res.status(200).json({ deleted_count: status.deletedCount, db_ok: status.ok === 1 });
 };
-
-/*
-exports.post_dialogs = function (req, res) {
-    let params = req.body;
-    if (!params.name || !params.intent || !params.intent.name || !params.intent.responses || params.intent.responses.length === 0) {
-        res.status(400).json({ "error": "Request needs a name (string), and an intent which contains a name & an array of responses." })
-    }
-    let insert = db.insertDialog(params);
-    res.json({ response: insert });
-};
-
-exports.get_dialogs = function (req, res) {
-    db.DialogModel.find({}, function (err, data) {
-        if (err) res.status(400).json({ "error": true, "message": err });
-        res.json(data);
-    });
-};
-*/
-
-exports.get_intents_analytics = async function (req, res) {
-    var lastWeek = new Date();
-    lastWeek.setDate(lastWeek.getDate() - 7);
-    let dbQueryParam = { "last_updated": { "$gt": lastWeek, "$lt": Date.now() } };
-    try {
-        var dates = await db.IntentsModel.find(dbQueryParam).select('last_updated');
-    } catch (err) {
-        res.status(400).send({ "error": true, "message": err });
-        return;
-    }
-    res.status(200).json(fromDbToFromattedAnalytics(dates));
-    return;
-}
-
-exports.get_entities_analytics = async function (req, res) {
-    var lastWeek = new Date();
-    lastWeek.setDate(lastWeek.getDate() - 7);
-    let dbQueryParam = { "last_updated": { "$gt": lastWeek, "$lt": Date.now() } };
-    try {
-        var dates = await db.EntitiesModel.find(dbQueryParam).select('last_updated');
-    } catch (err) {
-        res.status(400).send({ "error": true, "message": err });
-        return;
-    }
-    res.status(200).json(fromDbToFromattedAnalytics(dates));
-    return;
-}
-
-function fromDbToFromattedAnalytics(data) {
-    if (!data) {
-        return null;
-    }
-    let values = {
-        "Monday": {
-        },
-        "Tuesday": {
-        },
-        "Wednesday": {
-        },
-        "Thursday": {
-        },
-        "Friday": {
-        },
-        "Saturday": {
-        },
-        "Sunday": {
-        },
-    };
-    for (const date of data) {
-        let parsedDate = new Date(date.last_updated);
-        let dayAsString = dayOfWeekAsString(parsedDate.getDay());
-        let hoursAsString = parsedDate.getHours().toString();
-        if (values[dayAsString][hoursAsString] == null) {
-            values[dayAsString][hoursAsString] = 1;
-        } else {
-            values[dayAsString][hoursAsString]++;
-        }
-    }
-
-    let arr = [];
-
-    for (const item in values) {
-        for (const child in values[item]) {
-            arr.push({
-                name: item,
-                value: parseInt(child),
-                count: values[item][child],
-            })
-        }
-    }
-    return arr;
-}
-function dayOfWeekAsString(dayIndex) {
-    return ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"][dayIndex];
-}
-
-exports.generate_intents = async function (req, res) {
-    const MINS_IN_WEEK = 10080;
-    for (let i = 0; i < 100; i++) {
-        let name = uuid();
-        var date1 = new Date();
-        date1.setMinutes(date1.getMinutes() - getRandomInt(MINS_IN_WEEK));
-        let intent = new db.IntentsModel({
-            name: name,
-            expressions: [uuid(), uuid(), uuid()],
-            last_updated: date1,
-        });
-        const newIntent = await intent.save();
-        if (newIntent !== intent) {
-            res.status(400).json({ "error": true, "message": "Error when saving to db.", "details": newIntent });
-            return;
-        }
-    }
-
-    res.status(200).json({ "message": "populated db with intents" });
-}
-
-exports.generate_entities = async function (req, res) {
-    const MINS_IN_WEEK = 10080;
-    for (let i = 0; i < 100; i++) {
-        let name = uuid();
-        var date1 = new Date();
-        date1.setMinutes(date1.getMinutes() - getRandomInt(MINS_IN_WEEK));
-        let intent = new db.IntentsModel({
-            name: name,
-            expressions: [uuid(), uuid(), uuid()],
-            last_updated: date1,
-        });
-        let entity = new db.EntitiesModel({
-            name: uuid(),
-            synonyms: {
-                synonym_reference: uuid(),
-                list: [
-                    uuid(), uuid(), uuid(),
-                ],
-            },
-            last_updated: date1,
-        });
-        const newEntity = await entity.save();
-        if (newEntity !== entity) {
-            res.status(400).json({ "error": true, "message": "Error when saving to db.", "details": newEntity });
-            return;
-        }
-    }
-
-    res.status(200).json({ "message": "populated db with entities" });
-}
-
-function uuid() {
-    // generate a short uuid
-    return '_' + Math.random().toString(36).substr(2, 15);
-};
-function getRandomInt(max) {
-    return Math.floor(Math.random() * Math.floor(max));
-}
