+++ server/server.js
@@ -2,6 +2,7 @@ const express = require('express')
 const cors = require('cors')
 var router = express.Router();
 var db_api_controller = require('./Api/db_api');
+var db_helper_api_controller = require('./Api/db_helper_api');
 var converter_api_controller = require('./Api/converter_api');
 require('dotenv').config()
 
@@ -26,23 +27,24 @@ app.get('/', (req, res) => {
 });
 
 router.post('/intents', db_api_controller.post_intents);
-router.post('/intents-generation', db_api_controller.generate_intents);
+router.post('/intents-generation', db_helper_api_controller.generate_intents);
 router.post('/entities', db_api_controller.post_entities);
-router.post('/entities-generation', db_api_controller.generate_entities);
+router.post('/entities-generation', db_helper_api_controller.generate_entities);
 // router.post('/api/dialogs', db_api_controller.post_dialogs);
 router.post('/upload-csv', converter_api_controller.post_upload_csv);
 
 router.get('/intents', db_api_controller.get_intents);
-router.get('/intents-analytics', db_api_controller.get_intents_analytics);
+router.get('/intents-analytics', db_helper_api_controller.get_intents_analytics);
 router.get('/intents/:name', db_api_controller.get_intents);
 router.get('/entities', db_api_controller.get_entities);
-router.get('/entities-analytics', db_api_controller.get_entities_analytics);
+router.get('/entities-analytics', db_helper_api_controller.get_entities_analytics);
 router.get('/entities/:name', db_api_controller.get_entities);
 // router.get('/api/dialogs', db_api_controller.get_dialogs);
 router.get('/download-md/:filename', converter_api_controller.get_download_md);
 
 router.delete('/intents/:name', db_api_controller.remove_intent);
 router.delete('/entities/:name', db_api_controller.remove_entity);
+router.delete('/clear-database', db_helper_api_controller.clear_db);
 
 app.use('/api', router);
